+++ projects/net.splitcells.dem.merger/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 ## [Unreleased]
 ### Major Changes
+* **2022-11-07** Remove unused methods `CommonFunctions#findSystemOutput`.
 * **2022-04-03** **\#10**: Rename `Resource` interface to `ResourceOption` and it's `ResourceI`
      implementation to `ResourceOptionI` in order to clarify its meaning.
      This also makes it possible to use the name `Resource` for an interface